Rename generic JdbcDriver to FallbackJdbcDriver
People keep using JdbcDriver when they shouldn't. This name change better convey's it's idea in the name.

Examples of people misusing it:
http://stackoverflow.com/questions/24000759/scala-slick-delete-not-working
http://stackoverflow.com/questions/28529005/exception-when-upgrading-to-slick-3-0-0-m1-asynchronous-query
