+++ slick-testkit/src/test/scala/scala/slick/test/compile/NestedShapesTest.scala
@@ -1,7 +1,7 @@
 package scala.slick.test.compile
 
 import com.typesafe.slick.testkit.util.ShouldNotTypecheck
-import scala.slick.driver.JdbcDriver.api._
+import scala.slick.driver.FallbackJdbcDriver.api._
 import scala.slick.lifted.{Shape, ShapeLevel}
 
 class NestedShapesTest {
