+++ src/main/scala/scala/slick/jdbc/TypedStaticQuery.scala
@@ -372,7 +372,7 @@ object TypedStaticQuery {
 //    TODO: Fix this!
 //    lazy final val JdbcDriver = Class.forName(connectionParameter(jdbcDriver, "jdbcDriver")).asInstanceOf[java.sql.Driver]
     
-//    lazy final val SlickDriver = Class.forName(connectionParameter(slickDriver, "slickDriver")).asInstanceOf[scala.slick.driver.JdbcDriver]
+//    lazy final val SlickDriver = Class.forName(connectionParameter(slickDriver, "slickDriver")).asInstanceOf[scala.slick.driver.FallbackJdbcDriver]
 
   }
 }
