+++ src/sphinx/userdefined.rst
@@ -56,7 +56,7 @@ common scenario is mapping an application-specific type to an already supported 
 This can be done much simpler by using
 :api:`MappedColumnType <scala.slick.driver.JdbcProfile@MappedColumnType:JdbcDriver.MappedJdbcType.type>`
 which takes care of all the boilerplate. It comes with the usual import from the driver. Do not import
-it from the :api:`JdbcDriver <scala.slick.driver.JdbcDriver$>` singleton object.
+it from the :api:`FallbackJdbcDriver <scala.slick.driver.FallbackJdbcDriver$>` singleton object.
 
 .. includecode:: code/LiftedEmbedding.scala#mappedtype1
